refactor: split runner parallel helpers into package
